improve component structure
