+++ src/client/components/elements/Button/components/CenterButton.tsx
@@ -1,7 +1,7 @@
 import { useContext } from 'react'
-import Button from './Button'
+import Button from '../Button'
 import styled from 'styled-components'
-import { GameContext } from '../../../session/Context'
+import { GameContext } from '../../../../session/Context'
 import { useNavigate } from 'react-router-dom'
 
 const CenterButtonWrapper = styled.div`
