+++ src/client/components/elements/Button/components/SelfAlignedButton.tsx
@@ -1,4 +1,4 @@
-import Button from './Button'
+import Button from '../Button'
 import styled from 'styled-components'
 
 const SelfAlignedButton = styled(Button)`
