+++ src/client/components/elements/Container/components/GameAnswerContainer.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import Container from './Container'
+import Container from '../Container'
 
 const GameAnswerContainer = styled(Container)`
   display: flex;
