+++ src/client/components/elements/Container/components/LayoutContainer.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import Container from './Container'
+import Container from '../Container'
 
 const LayoutContainer = styled(Container)`
   min-height: 100vh;
