+++ src/client/components/elements/Container/components/TableContentContainer.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import Container from './Container'
+import Container from '../Container'
 
 const TableContentContainer = styled(Container)`
   display: grid;
