+++ src/client/components/elements/Input/components/InputCardBack.tsx
@@ -1,4 +1,4 @@
-import Input from './Input'
+import Input from '../Input'
 import styled from 'styled-components'
 
 const InputCardBack = styled(Input)`
