+++ src/client/components/elements/Input/components/InputCardFront.tsx
@@ -1,4 +1,4 @@
-import Input from './Input'
+import Input from '../Input'
 import styled from 'styled-components'
 
 const InputCardFront = styled(Input)`
