+++ src/client/components/elements/Label/components/ScoreLabel.tsx
@@ -1,4 +1,4 @@
-import Label from './Label'
+import Label from '../Label'
 import styled from 'styled-components'
 
 const ScoreLabel = styled(Label)`
