+++ src/client/components/handlers/CardRows/CardRows.tsx
@@ -1,7 +1,7 @@
 import NoResultsMessage from '../../typography/texts/NoResultsMessage'
 import Button from '../../elements/Button/Button'
 import RouterLink from '../../typography/links/RouterLink'
-import TableContentContainer from '../../elements/Container/TableContentContainer'
+import TableContentContainer from '../../elements/Container/components/TableContentContainer'
 import { CardProps } from '../../../../shared/types'
 import CardItem from '../../elements/CardItem/CardItem'
