+++ src/client/components/handlers/SortHeader/SortHeader.tsx
@@ -1,4 +1,4 @@
-import TableHeaderContainer from '../../elements/Container/TableHeaderContainer'
+import TableHeaderContainer from '../../elements/Container/components/TableHeaderContainer'
 import SimpleTableHeader from '../../typography/headings/SimpleTableHeader'
 import Label from '../../elements/Label/Label'
 import { InputType, SortDirection } from '../../../common/types'
