+++ src/client/components/layouts/AppBar/Appbar.tsx
@@ -1,6 +1,6 @@
-import CenterButton from '../../elements/Button/CenterButton'
+import CenterButton from '../../elements/Button/components/CenterButton'
 import TitleHeader from '../../typography/headings/TitleHeader'
-import TopBannerContainer from '../../elements/Container/TopBannerContainer'
+import TopBannerContainer from '../../elements/Container/components/TopBannerContainer'
 import TopBannerRouterLink from '../../typography/links/TopBannerRouterLink'
 
 export default function Appbar() {
