+++ src/client/components/layouts/GamePage/GameResultPage.tsx
@@ -1,10 +1,10 @@
 import CenterHeader from '../../typography/headings/CenterHeader'
 import FilledTableHeader from '../../typography/headings/FilledTableHeader'
 import Item from '../../typography/texts/Item'
-import ScoreLabel from '../../elements/Label/ScoreLabel'
-import GameContainer from '../../elements/Container/GameContainer'
-import GameResultsContainer from '../../elements/Container/GameResultsContainer'
-import StartButton from '../../elements/Button/StartButton'
+import ScoreLabel from '../../elements/Label/components/ScoreLabel'
+import GameContainer from '../../elements/Container/components/GameContainer'
+import GameResultsContainer from '../../elements/Container/components/GameResultsContainer'
+import StartButton from '../../elements/Button/components/StartButton'
 import { Fragment, useContext } from 'react'
 import { GameContext } from '../../../session/Context'
 import { startNewGame } from '../../../session/helper'
