+++ src/client/components/layouts/GamePage/OngoingGamePage.tsx
@@ -1,10 +1,10 @@
 import { useState, useContext, useEffect } from 'react'
-import GameButton from '../../elements/Button/GameButton'
+import GameButton from '../../elements/Button/components/GameButton'
 import ProgressHeader from '../../typography/headings/ProgressHeader'
-import ProgressLabel from '../../elements/Label/ProgressLabel'
-import GameAnswerContainer from '../../elements/Container/GameAnswerContainer'
-import GameContainer from '../../elements/Container/GameContainer'
-import InputAnswer from '../../elements/Input/InputAnswer'
+import ProgressLabel from '../../elements/Label/components/ProgressLabel'
+import GameAnswerContainer from '../../elements/Container/components/GameAnswerContainer'
+import GameContainer from '../../elements/Container/components/GameContainer'
+import InputAnswer from '../../elements/Input/components/InputAnswer'
 import Flashcard from '../../elements/Flashcard/Flashcard'
 import { GameContext } from '../../../session/Context'
 import { GameState, startNewGame } from '../../../session/helper'
