+++ src/client/components/layouts/GamePage/StartNewGamePage.tsx
@@ -1,6 +1,6 @@
 import Paragraph from '../../typography/texts/Paragraph'
-import GameContainer from '../../elements/Container/GameContainer'
-import StartButton from '../../elements/Button/StartButton'
+import GameContainer from '../../elements/Container/components/GameContainer'
+import StartButton from '../../elements/Button/components/StartButton'
 import { GameContext } from '../../../session/Context'
 import { GameState, startNewGame } from '../../../session/helper'
 import { useContext } from 'react'
