+++ src/client/components/layouts/ManageCardDetails/CardDetailPage.tsx
@@ -1,9 +1,9 @@
-import CardBackLabel from '../../elements/Label/CardBackLabel'
-import CardFrontLabel from '../../elements/Label/CardFrontLabel'
-import CardDetailContainer from '../../elements/Container/CardDetailContainer'
-import InputCardFront from '../../elements/Input/InputCardFront'
-import InputCardBack from '../../elements/Input/InputCardBack'
-import SelfAlignedButton from '../../elements/Button/SelfAlignedButton'
+import CardBackLabel from '../../elements/Label/components/CardBackLabel'
+import CardFrontLabel from '../../elements/Label/components/CardFrontLabel'
+import CardDetailContainer from '../../elements/Container/components/CardDetailContainer'
+import InputCardFront from '../../elements/Input/components/InputCardFront'
+import InputCardBack from '../../elements/Input/components/InputCardBack'
+import SelfAlignedButton from '../../elements/Button/components/SelfAlignedButton'
 import { useEffect, useState, useContext } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import { fetchCards, updateCard } from '../../../api'
