+++ src/client/pages/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom'
 import Appbar from '../components/layouts/AppBar/Appbar'
-import LayoutContainer from '../components/elements/Container/LayoutContainer'
+import LayoutContainer from '../components/elements/Container/components/LayoutContainer'
 
 export default function Layout() {
   return (
